:tada: Runtime refactoring
